engine: extract dispatch::actions (#159)
